matrix diagnostics

+++ src/velocities.cpp
@@ -28,13 +28,13 @@ double find_velocity_derivative(const double &r, const int &p, const double &r0)
     // double numerator = (1./r0) * (static_cast<double>(p+1)*pow(1.-pow(r/r0, 2), p)*pow(r/r0, 2)-pow(1.-(1.-pow(r/r0, 2)), p+1));
     // double denominator = static_cast<double>(p+1)*pow(r/r0,3);
 
-    double numerator = (1./r0) * (static_cast<double>(p+1) * pow(1.-pow(r/r0,2), p) * pow(r/r0,2) - (1-pow(1-pow(r/r0, 2), p+1)));
-    double denominator = (p+1)*pow(r/r0, 3);
+    double numerator = (1./r0) * ((static_cast<double>(p+1) * pow(1.-pow(r/r0,2), p) * pow(r/r0,2)) - (1.-pow(1.-pow(r/r0, 2), p+1)));
+    double denominator = static_cast<double>(p+1)*pow(r/r0, 3);
     return numerator / denominator;
   }
   else if (r > r0)
   {
-    return -pow(r0,2)/(pow(r, 3)*static_cast<double>(p+1));
+    return -(pow(r0,2)/(pow(r, 3)*static_cast<double>(p+1)));
   }
   else if (r0==0.)
     return 0.;
